# Non-native field multiplication optimization 1