* Halve the number of plookups
    * Eliminate custom constraint for scaling carry1_hi
    * Remove unused CircuitGateErrors (n.b. clippy doesn't detect these)
    * Add native modulus constraint test case
    
Speedup:
- time to create prover index: 62.8/62s ~ 1.3%
- time to create proof: 36/27s ~ 33%
- time to verify: 870/837ms ~ 3.9% 

```text
OPT1

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 844ms

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 838ms

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 834ms

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 837ms

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 832ms

=== averaged over 5 runs ===

- time to create prover index: 62s
- time to create proof: 27s
- time to verify: 837ms
```

```text
VANILLA

- time to create prover index: 63s
- time to create proof: 36s
- time to verify: 831ms

- time to create prover index: 63s
- time to create proof: 36s
- time to verify: 866ms

- time to create prover index: 63s
- time to create proof: 36s
- time to verify: 837ms

- time to create prover index: 62s
- time to create proof: 36s
- time to verify: 977ms

- time to create prover index: 63s
- time to create proof: 36s
- time to verify: 839ms

=== averaged over 5 runs ===

- time to create prover index: 62.8s
- time to create proof: 36s
- time to verify: 870ms
```

Closes #923